Fix for #1804 - Alignment issues in Fortran with MPI_Status
Changed the Fortran wrapper generator to test for proper alignment of
MPI_Status - because of the new MPI_Count field, some systems may need
MPI_Status to be aligned on a multiple of sizeof(MPI_Count).  By
default, new code enforces that, using a temporay status variable only
when required.  Cost is a few additional tests.