Fixes for #1804 and #1828, related to alignment issues with MPI_Status
There are two alignment-related problems with MPI_STATUS arguments in
Fortran.  The first is that while MPI_STATUS is an INTEGER array in
Fortran, in C there is an element of type MPI_Count.  MPI_Count may
be longer than an INTEGER and thus a C MPI_Status may have a stricter
alignment requirement than a Fortran INTEGER array. The fixes to
src/binding/f77/buildiface ensures that STATUS arguments that are
provided to C by the Fortran wrappers are properly aligned, while avoiding
copies when possible.  The second fix deals with padding that the
introduction of MPI_Count as the count type into the C structure for
status will often introduce.