Remove caching wrapper for gethostname.
This was added for terrible systems that rely on a slow DNS server to
look up the local hostname.  Most OSs do a much better job setting
local host names in the local /etc/hosts file today.  Even if they
don't, we no longer do a large number of lookups anymore, so this
should not be a problem.