fix bug of zero-byte shared window allocation.
In MPIDI_CH3I_Win_allocate_shared, if the size of shared memory
segment is 0, we just set window's base to NULL, do not allocate
shared memory segment and mutex memory region.

In MPIDI_CH3I_SHM_Win_free, we detach the shared memory segment
and mutex memory region only when the size of shared memory segment
is not 0.

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>