resized wasn't setting max_contig_count
Not much in MPICH consults the max_contig_count field of the datatypes,
except ROMIO.  Resized types were reporting "huge negative number" for
the count of contiguous blocks because datatype code wasn't setting the
max_contig_count member.  Why 3?  The lower bound and upper bound parts
of the datatype, while not actually data or blocks, are effectively
blocks for the purposes of the one client of this code -- the ROMIO
datatype flattening code.

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>