Recent versions of Cray Fortran may require a special arg to create modules
While testing on the Blue Waters development system, builds failed
because the code for determining the Fortran Module options was unable to
figure out the Cray compiler (naturally, this used to work).  Investigation
showed that the compiler needs a "-em" option to create separate module
files.  This change adds a test for that, if the compiler is unable to
create module files without that option.  This also cleans up several other
minor problems with this code.