Fixed touch command option order
Fix touch command in autogen.sh that generates a random file and reruns
configure during first make.

Some systems (MACs) require the options to be specified prior to the target
file for the touch command. This commit fixes the option order in the touch
command in autogen.sh introduced to fix ticket #1870.

Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>