Error in the buffer attached to dynamic windows.
In MPICH, Win_attach/detach are essentially no-ops.  The actual
accumulate calls don't actually check if the actual buffer is attached
or not.  This is technically a user error.  The win_dynamic_acc test
had a wrong buffer attached, but was still passing because of this
reason.  This patch only corrects the test program.

Thanks to Michael Raymond @ SGI for reporting the error.

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>