When judging if origin and target process are on the same node, using vc->node_id flag instead of vc->ch.is_local flag.
Flag 'is_local' is not correct because it is defined in nemesis, not in CH3.
Flag 'node_id' is defined in CH3.

Note that for ch3:sock, even if origin and target are on the same node, they are not within the same SHM region.
Currently ch3:sock is filtered out by checking shm_allocated flag first. In future we need to figure out a way to
check if origin and target are within the same "SHM comm".

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>