Clean up pamid MPID_Abort() logic
Removed the processing of the (undocumented) environment variable
'PAMID_CORE_ON_ABORT' which was being checked to determine if the user
does *not* want the process to core dump. On Blue Gene/Q the core dump
was accomplished by calling 'abort()' which sends SIGSBRT to all
processes and all processes would then write a core file. This is not
scalable.

Instead, MPID_Abort() will invoke 'exit(1)' which will terminate all
processes in the job. This behavior is identical for both the POE and
the Blue Gene/Q control systems.

On Blue Gene/Q the user may replicate the previous core dump behavior by
using the environment variables 'BG_COREDUMPONERROR=1' or
'BG_COREDUMPONEXIT=1'.

Finally, the 'DYNAMIC_TASKING' #ifdef is moved up so it is checked first.
'MPIDI_NO_ASSERT' and 'DYNAMIC_TASKING' are typically defined for PE. It
appears that the dynamic tasking code was never being invoked.

(ibm) CPS 99YURA

Signed-off-by: Bob Cernohous <bobc@us.ibm.com>