Fix pamid bgq configure for v1r2m1 installs
There are three configure options and two install cases that need to be
handled.

The configure could have be done on a bgq system that has V1R2M1+
installed or on a bgq system that has a pre-V1R2M1 installation. In
V1R2M1 the location and names of the pami libraries changed.

The three configure options are:
  --with-bgq-install-dir ..... forces use of a specific bgq install
  --with-pami[-include|-lib] . forces use of a specific pami install
  PAMILIBNAME ................ forces use of a specific pami library name

Signed-off-by: Bob Cernohous <bobc@us.ibm.com>
Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>