MPI_Status updates for ABI.
Based on an Intel contributed patch.  The idea is to use the bits from
the cancelled field to extend the count, rather than increasing the
count datatype itself.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>

Fixes to the bit manipulation based on feedback from Artem Yalozo @
Intel.

Fixes to the naming convention based on feedback from Bill Gropp.

Signed-off-by: William Gropp <wgropp@illinois.edu>