Configure check for (need for) header padding
Dancing around with various schemes to detect/deal-with compiler padding
in structs so the 'payload' field ends up double-aligned.  The new
"large type , large count" work introduced a pointer into the nemisis
packet, so we have to be a bit more dynamic about how or if we pad out
the structure.

- put the members of MPID_nem_pkt_header inside a struct, so subsequent
  payload does not "fall into" compiler-introduced padding
- use this updated datatype to determine the size, and if we need to pad
  or not.
- use the result of the configure test to introduce padding, or omit the
  padding from the definition if the amount of padding is '0'

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>