Check for BUILTIN kinds, not just predefined datatypes.
Some builtin kinds (e.g., MPI_DOUBLE_INT), can have extra padding when
a vector of them is copied.  So we need to use the extent for such
types, not just the size.  In this commit, we just simplify the
optimization to only work on BUILTIN types, which do not have this
issue.

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>