Move the errors/f77/errhan directory to an iodir.
The test file was using MPI-I/O functions that the MPICH test suite
allows us to enable/disable.  This patch moves the test to an iodir
format, so it is not run when configured with --disable-romio.

The correct longer term fix for this is to allow multiple conditions
to disable a given test.  For example, we could have configure set the
@enable_io@ macro to either '' or '#', and different tests that rely
on I/O would add them to the start of the line in the test file.  That
way, a test that uses both F77 and I/O would be disabled when support
for either one is not present.

Signed-off-by: Junchao Zhang <jczhang@mcs.anl.gov>