Hotfix for 53640e26
The commented-out variable in 53640e26 is in fact hiddenly used by the macro
MPIU_ERR_SET1, only if HAVE_ERROR_CHECKING is defined. That macro assumes the
variable is declared and does not take it from its arguments. This hotfix
declares the FCNAME[] variable if HAVE_ERROR_CHECKING is defined.