Added more MPICH authors and removed old ones.
Changes from Toni:

Modifications in documentation author formatting

1.- Modified the Spanish weird n-with-a-nice-tilde-on-top of Toni's
last name to look as a regular letter instead of an even weirder thing.
2.- Included Toni's middle initial to match his bibliographic name.

Removed "and" from authors on docs

As the template introduces space among authors, it looks weird to me
that the last author gets an "and" word in front of it.

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>