compile wrapper cleanup
Simplify logic in compile wrapper scripts. Use configure substitutions
where possible to better match pkg-config style.

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>

Includes the following modifications by Pavan Balaji:

Remove the PAC_COMPILER_SHLIB_FLAGS usage, instead of modifying the
macro in confdb.

The ordering of flags in mpicc and friends does not match that of
pkg-config.  This is because of two reasons.

1. pkg-config reorders flags when it outputs them.  This requires us
to manually adjust the flags in mpicc to match up, and is error prone.

2. mpicc and friends provide LDFLAGS before the user-specified flags,
followed by the include and library directories.  This is to make sure
that the LDFLAGS are listed before the application source file.
Reordering them to match pkg-config loses this flexibility.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>