Make SMP-awareness in MPI_Barrier consistent.
The current implementation of SMP-awareness in MPI_Barrier was in
MPIR_Barrier_impl.  This makes the default implementation of barrier
SMP-aware.  However, if a device overrides barrier and then calls back
the default implementation through MPIR_Barrier, it can no longer take
advantage of SMP-awareness.  This patch moves the SMP-aware
implementation to MPIR_Barrier_intra, which is called by
MPIR_Barrier_impl through MPIR_Barrier, in the default implementation.
See #1957.

Signed-off-by: Michael Blocksome <blocksom@us.ibm.com>