Thread critical-section initialization fixes.
This patch has two related parts.

1. We initialize the isThreaded runtime check to FALSE before doing
the cvar checks.  This is because the cvar initialization uses memory
allocation and string duping, which internally use mutexes.  But the
mutexes are not initialized yet, so disabling isThreaded would make
sure we don't use them.

2. Updates to the pamid device to respect the isThreaded variable.

This is needed since we were initializing cvars before setting the
thread-level, so the thread-safety macros are not initialized at that
point.  See #1900.

Signed-off-by: Michael Blocksome <blocksom@us.ibm.com>
Signed-off-by: Rob Latham <robl@mcs.anl.gov>