Prioritize thread-multiple branch.
Give a "likely" attribute to the branch that checks if an MPI process
is initialized in THREAD_MULTIPLE or not.  This should reduce the cost
of the branch for applications that require THREAD_MULTIPLE.

The per-object mode in the pamid branch is intended for the case where
comm-threads are used.  In this case, the application is mostly in
THREAD_MULTIPLE mode, even if it didn't explicitly initialize it as
such.  The only time it'll not be in that mode is during
initialization before the appropriate mutexes have been setup.

See #1900.

Signed-off-by: Michael Blocksome <blocksom@us.ibm.com>
Signed-off-by: Rob Latham <robl@mcs.anl.gov>