Comment out nb_test, since it's not entirely correct.
This was meant to test out the case when MPI_Test is not nonblocking.
However, we ended up assuming that MPI_Win_lock will be nonblocking.
That is not specified by the standard and might not be true.
Commenting this out till be find a better way to test the original
problem with MPI_Test.

Fixes #1910.

Signed-off-by: Rajeev Thakur <thakur@mcs.anl.gov>