Add -lgcov to LIBS when --enable-coverage
When mpich is configured with ./configure --enable-coverage ..., we put
-fprofile-arcs -ftest-coverage to CXXFLAGS. According to gcc manual,
-fprofile-arcs -ftest-coverage can be replaced by --coverage. But unfortunately,
I found though both add -lgcov when linking libmpichcxx.so, -lgcov is put to the
end of the link command, which is too late.

Adding -lgcov directly to LIBS will put -lgcov before -lc, which makes
the linker correctly resolve symbol 'atexit' in linking.

Fixes #2000

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>