Add on-demand connection management for IB
A pair of QPs connecting two MPI processes is created only when a
communication between them is requested.
Allocation of eager-send ring-buffer is implemented as well. First N
connctions are given ring-buffers used only for them and the following
connections are given a shared ring-buffer, similar to send/recv, of
which a sender needs to reserve a buffer slot before initiating
RDMA-write.

Signed-off-by: Pavan Balaji <balaji@anl.gov>