Adds universe check to C spawn tests
The test suite assumes that it is running on a system where
MPI_COMM_SPAWN(_MULTIPLE) is possible. This isn't true on all systems and thus
prevents the test suite from being useful everywhere.

This check (contributed by Jeff H. - jhammond@alcf.anl.gov) looks to see if
the MPI_COMM_WORLD attribute MPI_UNIVERSE_SIZE is available and actually
greater than the size of MPI_COMM_WORLD.

If the spawn tests will fail because spawning is not possible on the system
(whether because the universe size is too small or spawning is not supported),
the tests will return success (unless there was some error in the detection
process).

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>