Moves the tag reservation to MPI layer
Resets MPIDI_TAG_UB back to 0x7fffffff. This value was changed a while back,
but the change should have happened at the MPI layer instead of the CH3 layer.
This resets the value to allow CH3 to use the tag space.

Instead, the value is now set in the MPI layer during initthread. This means
that it will be safe regardless of the device being used. This prevents a
collision that was occurring on the pamid device where the values for
MPIR_TAG_ERROR_BIT and the MPIR_Process.attr.tagged_coll_mask values were the
same.

Fixes #2008

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>