Add sendrecv2 and sendrecv4 test from MPICH-1
This test moves the sendrecv2 and sendrecv4 test from MPICH-1 to the new
MPICH tests.  It has been updated to use the MTest framework.  This test
uses the datatypes created by the dtypes.c file (also moved from
MPICH-1).

These tests involve creating a structure type with an explicit offset of
a variety of types, and then communicated with MPI_BOTTOM as the buffer
location.  Several of these tests failed in a recent update to the Dataloop
code for optimization of performance.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>