Move communicator destruction to after progress checks.
During finalize, we were destroying the COMM_WORLD, COMM_SELF and
COMM_IWORLD communicator objects, and all other associated resources
internally, before waiting for the final progress checks for incoming
messages finished.  This resulted in the following sequence of cleanup:

1. COMM_WORLD got cleaned up.  Internally, there is a check to see if
a group object has been allocated for COMM_WORLD.  If there is one, it
is freed up.

2. We waited for other messages to arrive.  We noticed a failure at
this time, so we try to create a failed process group.  This uses the
COMM_WORLD group internally, causing it to be created again, but with
a reference count of 2, since the code assumes that the first
reference count is always for the original COMM_WORLD.

3. When we try to free the world group, we notice that the reference
count is 2, so we decrement the reference count and not actually free
the object.

Moving the check for incoming messages to happen before the
communicator free, fixes this problem.  Note that the PG finalization
still needs to be the last step since that cleans up all the VCs as
well.

See #1996

Signed-off-by: Wesley Bland <wbland@mcs.anl.gov>