Improve PMI_process_mapping parsing.
The original PMI process mapping parsing code had a number of
assumptions that would allow it to only work on COMM_WORLD.  This
patch corrects these to work for dynamic processes as well.

It also corrects the evaluation of the number of nodes used to be
correct in the general case.

Refs #2007.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>