Bug-fix: deal with wrap around PMI process mapping strings
If the PMI process mapping string wraps around to node 0, we were
creating a bad node list of which processes are local and which are
not.  This patch provides a hacky fix for this case by only repeating
the part of the PMI mapping string from the point where it wrapped
around.

The patch is hacky because it assumes that seeing a start node ID of 0
means a wrap around.  This is not necessarily true.  A user-defined
node list can use the node ID 0 without actually creating a wrap
around.  The reason this patch still works in this case is because
Hydra creates a new node list starting from node ID 0 for
user-specified nodes during MPI_Comm_spawn{_multiple}.  If a different
process manager searches for allocated nodes in the user-specified
list, this patch will break.

Fixes #2007.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>