Reduce the number of processes needed by rma-contig.
This test does not do anything special for 4 processes.  rank 0
communicates with each other process sequentially, and waits for all
communication to be over before communicating with a different target.
Reducing the number of processes allows us to reduce the time taken by
the test without affecting any tested parameter.

Signed-off-by: Wesley Bland <wbland@mcs.anl.gov>