Change the pamid default for advancing the scheduled NBCs
In the pamid layer, mpir_nbc is set to 0 as the default e.g. the scheduled NBC
work will not be advanced which would cause the job hang. The NBC would work
only if the schedule is advanced. To do so, a user requires to set the
environment variable PAMID_MPIR_NBC to 1.

With MPICH 3.1 being released and NBC is one of support items in the release,
the default for mpir_nbc should be changed from 0 to 1.

Signed-off-by: Michael Blocksome <blocksom@us.ibm.com>
Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>