Improve localhost detection.
When compiled with strict we completely disable local host detection.
In the case where no host is specified, we use the local host name
that we get using gethostname.  However, when we try to detect if that
is the local host, we do not even attempt to do a simple gethostname
check.

Thanks to Orion Poplawski <orion@cora.nwra.com> for reporting the
problem and a detailed analysis of the issue.

Signed-off-by: Wesley Bland <wbland@mcs.anl.gov>