Remove overly conservative assertion.
MPI_COMM_IDUP currently just asserts out when we use it in
multithreaded mode.  This is overly conservative.  MPI_COMM_IDUP works
fine in such modes unless it is executed concurrently with another
routine that also tries to allocate or free a context ID.  Removing
this assertion till we find a more appropriate way of tracking this
error.

Fixes #2020.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>