Reorder C++ virtual functions to match Intel MPI.
This patch breaks ABI compatibility between this and older versions of
MPICH, but is needed to get MPICH closer to the ABI compatibility
initiative with Intel, IBM and Cray.

This patch does not make the resultant mpicxx.h exactly the same as
what Intel MPI ships, but most of the publicly exposed virtual
functions should be in order.  The remaining functions that are
out-of-order are Reduce_scatter_block and Call_errhandler for
communicators, which are placed after the internal Clone function.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>