Address many of the perf problems in #1788
This adds a step that optimizes the dataloop representation, primarily
merging CONTIG blocks with the parent dataloop, such as a VECTOR loop.
It also performs a change of VECTOR of CONTIG with extent > size to
VECTOR of VECTOR; this reduces the stack operations needed to perform
the move.

This is a temporary fix for the dataloop performance.  See the DAME
wiki page (http://wiki.mpich.org/mpich/index.php/DAME) for current work
on a replacement, higher performance datatype system.

A partial, but not complete, fix for ticket #1788

Signed-off-by: Rob Latham <robl@mcs.anl.gov>