Rename Fortran binding directories.
f77 and f90 were not consistent names with respect to what the code
was providing.  For example, "f90" also used code from "f77".  This
patch changes the naming to "mpif_h" and "use_mpi", which correspond
to using the mpif.h and "use mpi module" conventions specified by the
MPI standard.

This also includes changes to buildiface, autogen.sh and .gitignore to
work with these changes.

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>