Significant simplification of ad_bg_open
In order to accomodate deferred open, we can't do *any* collective
operations in ad_bg_open.  Any collectives have to happen one level up
at ADIOI_GEN_Opencoll.

We already promoted fs blksize in a prior patch, and simplified
"scalable sync" in another patch, so when we remove the collective call
(bcast of blocksize and fs type), we can also remove the "is it ok to
scalalbe sync"? (because it will always be ok) and the "are we an
fsync-aggregator" logic becuase now only the first io aggregator will be
such an aggregator.