fix $wl setting for Solaris Studio compilers
The Solaris/Sun compilers accept the -Wl, style flags for passing
arguments to the linker in version 8.5 on Linux. We wse this method,
as the compiler no longer passes unrecognized flags to the linker
by default. The matching logic in config.rpath is mostly borrowed
from libtool commit [a55208d5]. Fixes #2045

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>