Remove the use of MPIDI_TAG_UB
The constant MPIDI_TAG_UB is used in only one place at the moment, in the
initialization of ch3 (source:src/mpid/ch3/src/mpid_init.c@4b35902a#L131). The
problem is that the value which is being set (MPIR_Process.attrs.tag_ub) is
set differently in pamid (INT_MAX). This leads to weird results when we set
apart a bit in the tag space for failure propagation in non-blocking
collectives (see #2008).

Since this value isn't being referenced anywhere else, there doesn't seem to
be a use for it and it's just leading to confusion. To avoid this, here we
remove this value and just set MPIR_Process.attrs.tag_ub to INT_MAX in both
ch3 and pamid.

See #2009

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>