Rename mpich libraries.
The following library names are used to make the naming consistent
across the ABI compatibility group:

C libraries: libmpi.* and libpmpi.*
C++ library: libmpicxx.*
F77 libraries: libmpif77.*
F90+ library: libmpifort.*

This patch also gets rid of the FWRAPNAME variable, which is a
duplicate of MPIFLIBNAME.  Similarly, FCWRAPNAME is removed and a new
variable MPIFCLIBNAME is added, so it's consistent with the other
names.

PMPIFLIBNAME, which was unused, is no longer present.

Fixes #2039.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>