Move symbols to correct libraries.
Maintain a list of files that go into each library.  If a particular
binding is not enabled, the list variable still exists, but will just
be empty.  This simplifies the management of which files/symbols go
into which library.

Move all MPI_ symbols to the libmpi library and all other symbols to
the libpmpi library.  All Fortran 77 symbols go into libmpif77.so,
while C symbols go into libmpi.so.  There are some exceptions, such as
status_f2c, which are handled by the Fortran code but used in C.  Our
Fortran 90 build only creates a few symbols and uses the f77 symbols
for everything else.  These few symbols go into libmpifort.so.

Also update compiler wrappers to link to correct libraries.  mpif77
should now link with libmpif77.  mpif90 links with both libmpifort and
libmpif77, since our F90 build still keeps the core Fortran library
symbols in libf77.

We completely ignored the F77 library earlier.  This was OK because
all of the Fortran symbols were ending up in libmpi.  Now that we have
separated out the symbols to the right library, we now need to link to
libmpif77 as well.

Also added inter-library dependencies.

libmpi has a dependency on several internal libraries: libmpl, libopa.
libmpicxx did not have a dependency on libmpi, added.
libmpif77 did not have a dependency on libmpi, added.
libmpifort did not have a dependency on libmpi, added.

This dependency model is sufficient for C and F77, but not for C++ and
F90.  The C and F77 libraries contain all the symbols the application
relies on, but the F90 and C++ libraries don't.  In the case of F90,
symbols such as mpi_bcast are missing and are borrowed from the F77
library.  In the case of C++, mpicxx.h contains calls directly to C
functions (such as MPI_Reduce_local), which get embedded into the
application.

Fixes #2023.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>