Use library dependencies.
When the conditions are right, we use libtool's interlibrary
dependencies to only link the application with the minimum set of
libraries required.  Those libraries will pull in the other required
libraries on demand.  That way, the application library dependencies
would be minimal, allowing us to add/remove extra libraries in the
future.

The "conditions are not right" when one of the following is true:

1. MPICH shared library builds are disabled.

2. libtool doesn't support interlibrary dependencies on this platform.

3. The user passed -static or -mpich-alllibs to the command line.

4. Interlibrary dependency tracking is explicitly disabled at
configure time.

Fixes #2040.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>