modify MPL_trdump to make a single fprintf call
if there are a lot of mpi processes with memory errors, the output can
get horribly jumbled if the "memory at address" bit is emitted in one
fprintf call and the "at file/line" bit emitted in a second.  For each
memory region, Build up the address/file-name string in memory and make
one fprintf (instead of two) per memory region.

Signed-off-by: Pavan Balaji <balaji@anl.gov>