Fixes MPI_Comm_idup
This patch fixes two related tickets:
1. MPI_Comm_idup in multithreaded environments
2. MPI_Comm_idup fails to create multiple communicators
Because these two tickets are tightly coupled, so they are fixed in this
single patch.

The original code did not consider the multithreaded case and did not
use progress engine in correct order when saving a copy of global mask
to local thread.

Following changes were made to implement the MPI_Comm_idup correctly:

1. It shares the same global flag 'mask_in_use' with other communicator
functions to protect access to context_mask. And use CONTEXTID lock to
protext critical sections.

2. It uses the same algorithm as multithreaded MPI_Comm_dup
(multi-threaded vertion of MPIR_Get_contextid_sparse_group) to allocate
a context id, but in a nonblocking way. In the case of conflicts, the
algorithm needs to retry the allocation process again. In the
nonblocking algorithm, 1) new entries are inserted to the end of
schedule to replace the 'while' loop in MPI_Comm_dup algorithm; 2) all
arguments passed to sched_get_cid_nonblock are saved to gcn_state in
order to be called in the future; 3) in sched_cb_gcn_allocate_cid, if
the first try failed, it will insert sched_cb_gcn_copy_mask to the
schedule again.

3. There is a subtle difference between INTRACOMM and INTERCOMM when
duplicating a communicator.  They needed to be treated differently in
current algorithm. Specifically, 1) when calling sched_get_cid_nonblock,
the parameters are different; 2) updating newcommp->recvcontext_id in
MPIR_Get_intercomm_contextid_nonblock has been moved to
sched_cb_gcn_bcast because this should happen after
sched_cb_gcn_allocate_cid has succeed.

Fixes #1935
Fixes #1913

Signed-off-by: Pavan Balaji <balaji@anl.gov>