Add check for revoked communicator
Piggybacking on the MPID_Comm_valid_ptr check in the HAVE_ERROR_CHECKING
block, this checks to see if the communicator has been revoked and
returns MPIX_ERR_REVOKED if so.

This probably should move out of the HAVE_ERROR_CHECKING section since it
requires the user to have this turned on. If the user leaves it off, they'll
never be notified. However, if this moves out of the HAVE_ERROR_CHECKING
section, it will probably have performance implications.

Signed-off-by: Junchao Zhang <jczhang@mcs.anl.gov>