Add an explicit application dependency on the profiling library.
We were not explicitly linking the application binaries to libpmpi
earlier.  This works fine when the application uses the MPI_ symbols,
but not when the application uses the PMPI_ symbols.  On mac, where a
separate profiling library is built, we were using -flat_namespace to
supress this issue.  But on other platforms, without this flag, the
issue showed up when passing --disable-weak-symbols.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>