Make "F77" an expert-level setting.
The user can now enable/disable fortran using the --enable-fortran
option, instead of separate --enable-f77 and --enable-fc.  Expert
users can control specific Fortran standards to use with
--enable-fortran=f77,f90 (and f08 in the future).

Also allow the user to specify only "FC".  If "F77" is not specified,
we use the same compiler as FC.

We now use mpifort as the primary Fortran wrapper, with mpif90 as a
symlink to it for backward compatibility.  If the user specifies a
different F77 compiler than the FC compiler, or if FFLAGS and FCFLAGS
are different, then a separate mpif77 is generated.  Otherwise, mpif77
is a symlink to mpifort.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>