Remove support subarray in mpi_f08_constants.f90
The MPI-3 standard is broken in a bunch of ways with respect to
subarrays support.  For example, it is unclear as to which functions
can be passed a subarray and which cannot.  This is being fixed by the
WG.  In the meanwhile, we'll disable subarray types.

No review since F08 binding is experimental now.