pamid: mapped datatypes needed to be unmapped
for one-sided communication, the completion of a local operation is when
the completion handler of the operation is activated. The datatype
associated with the operation can not be freed until the completion
handler is activated which is MPIDI_Win_DoneCB. In the current code, the
mapped datatype is unmapped immediately freed after the routine gets the
control from PAMI calls. The calls could be PAMI_Send, PAMI_Get, PAMI_Rget,
PAMI_Put and PAMI_Rget. Please note that the datatype is needed for PAMI
to process the data.

The fix is to unmap the datatype in MPIDI_Win_DoneCB() and remove all
"unmap" datatype operations from put and get operations.

Signed-off-by: Michael Blocksome <blocksom@us.ibm.com>