pamid: memory leak on handles for request based RMA operations
Request based RMA communication operations require that each operation
allocates a communication request object and associates it with the request
handle (the argument request) that can be used to wait or test for
completion. Both the structures for the request handle and the window
structure are not being freed. The problem is when the request handle
is created, the ref_count of the object is set to 2. When
MPID_Request_releas_inline() is called by a request based RMA operation,
the ref_count of the handle is decremented to 1. The allocated memory spaces
are freed only if ref_count is equal to 0.

To fix the problem, set the ref_count to 1 when the request handle is
created. In MPID_Request_release_inline, the routine decrements the
ref_count to 0 which leads the memory space be freed.

Signed-off-by: Michael Blocksome <blocksom@us.ibm.com>