Separate progress engine code from garbage collection
Currently the code of poking progress engine to complete
requests and the code of cleanning up completed requests
are mixed up in one function rma_list_gc(), which is not
a clear code structure. We move the code of poking progress
engine out of rma_list_gc() and encapsule the code into
a separate function so that rma_list_gc() only does garbage
collection work.

Signed-off-by: Pavan Balaji <balaji@anl.gov>