ROMIO test case demonstrating indexed type bugs
The problem is when defining a filetype using MPI_Type_indexed and the
first few elements of argument blocklens[] are zeros, a collective write
will miss writing some data.

The test program first fills a file with 9 integers with values all
-999.  It then defines a filetype and writes to the file in parallel
with user buffers with value all 1s. Lastly, the file is read back and
checked for contents.

(it's Wei-keng's test case. I just hooked it into ROMIO's test suite)

Signed-off-by: Rob Latham <robl@mcs.anl.gov>