Replace malloc-free functions when using netmod-IB
We create malloc-free functions to reduce ibv_dereg_mr overhead.

We implement 'malloc' as follows. If there is a memory marked as usable,
we use it and mark as in-use. If not, we allocate memories and mark them
as usable. In the case of small size, we do mmap a large memory area and
divide it into some blocks.  In the case of big size, we do mmap by the
size aligned power-of-two.

We implement 'free' as follows. We check the free size by the address.
In the case of small size, we count the number of 'free' called. If all
blocks are freed, we mark them as usable. In the case of big size, we
mark it as usable.

Signed-off-by: Pavan Balaji <balaji@anl.gov>